renamed artifact-registry subproject
- now simply spring-cloud-dataflow-registry
- moved the URI registry related classes there as well
