+++ pom.xml
@@ -33,9 +33,9 @@
 	<modules>
 		<module>spring-cloud-dataflow-app-launcher</module>
 		<module>spring-cloud-dataflow-configuration-metadata</module>
-		<module>spring-cloud-dataflow-artifact-registry</module>
 		<module>spring-cloud-dataflow-deployer-spi</module>
 		<module>spring-cloud-dataflow-core</module>
+		<module>spring-cloud-dataflow-registry</module>
 		<module>spring-cloud-dataflow-rest-resource</module>
 		<module>spring-cloud-dataflow-server-core</module>
 		<module>spring-cloud-dataflow-server-local</module>
