--- spring-cloud-dataflow-artifact-registry/src/main/java/org/springframework/cloud/dataflow/artifact/registry/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Root package of the ArtifactRegistry support.
- */
-
-package org.springframework.cloud.dataflow.artifact.registry;
