+++ spring-cloud-dataflow-completion/pom.xml
@@ -14,7 +14,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-artifact-registry</artifactId>
+			<artifactId>spring-cloud-dataflow-registry</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
