+++ spring-cloud-dataflow-registry/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<artifactId>spring-cloud-dataflow-artifact-registry</artifactId>
+	<artifactId>spring-cloud-dataflow-registry</artifactId>
 	<packaging>jar</packaging>
 	<parent>
 		<groupId>org.springframework.cloud</groupId>
