+++ spring-cloud-dataflow-registry/src/main/java/org/springframework/cloud/dataflow/registry/AppRegistration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.dataflow.artifact.registry;
+package org.springframework.cloud.dataflow.registry;
 
 import java.net.URI;
