+++ spring-cloud-dataflow-registry/src/main/java/org/springframework/cloud/dataflow/registry/RedisUriRegistry.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.dataflow.server.registry;
+package org.springframework.cloud.dataflow.registry;
 
 import java.net.URI;
 import java.net.URISyntaxException;
