+++ spring-cloud-dataflow-registry/src/main/java/org/springframework/cloud/dataflow/registry/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Root package of the Data Flow registry support.
+ */
+
+package org.springframework.cloud.dataflow.registry;
