+++ spring-cloud-dataflow-server-core/pom.xml
@@ -30,7 +30,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-artifact-registry</artifactId>
+			<artifactId>spring-cloud-dataflow-registry</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
